feat: Update page and route for creating posts and the sidebar is working!!
Refactor the page.tsx file to update the main container's comment and fix a typo in the class name. Update the route in the constants/index.js file to reflect the change from "/create-thread" to "/create-post". Add a new file, create-post/page.tsx, to handle the creation of posts. This file includes a function component that displays the heading "Create Post". This commit improves the user experience by providing a clearer and more accurate route for creating posts.