feat: Refactor ThreadCard component to use flex layout and update styles
Refactor the ThreadCard component to use flex layout by adding the `flex` class to the `article` element. Update the styles by adding the `w-full`, `flex-col`, `rounded-xl`, `bg-lime-100`, and `p-7` classes to the `article` element. Also, update the `h2` element to use the `text-black` class instead of `text-small-regular text-light-2`. This improves the visual presentation of the ThreadCard component and enhances the user experience.